requirements/development.txt
-r base.txt

django-debug-toolbar==2.2
django-extensions==3.0.2
coverage==5.1
safety==1.9.0
Requirementz==0.3.5
pipdeptree==1.0.0
